refactor: use com.linroid.kdown.app.{platform} for source packages
Move source packages under com.linroid.kdown.app.{platform}:
- Android: com.linroid.kdown.app.android (namespace + source)
- Desktop: com.linroid.kdown.app.desktop (mainClass + source)

Application ID stays com.linroid.kdown.app for Android and iOS.
